⚡ Bolt: Move indexedCount calculation to DownloadsView
- Moved `indexedCount` calculation from `App.tsx` to `components/DownloadsView.tsx`.
- Updated `DownloadsView` to accept `syncedData` prop instead of `indexedCount`.
- Prevents unnecessary tree traversal when Downloads tab is not active.
